requirements.txt
PyQt5==5.15.9
python-dateutil==2.8.2
pillow==10.0.0  # For icon creation 